updated the start script, renamed the main file to main, and updated the start bot script

+++ startBot.sh
@@ -1,11 +1,18 @@
 #!/bin/bash
 
-deactivate  
+# Check if virtual environment is active
+if [[ -n "${VIRTUAL_ENV}" ]]; then
+    deactivate
+fi
 
-rm -rf venv
+if [ -d "venv" ]; then
+    rm -rf venv
+fi
 
 python -m venv venv
 
 source venv/bin/activate
 
 pip install -r requirements.txt
+
+python3 main.py
